+++ src/components/cardMaker/CardEditPage.tsx
@@ -1,7 +1,7 @@
 import useSelectedPokemonForCard from '@/store/useSelectedPokemonForCard';
 import SelectPokemonLike from '@/components/cardMaker/selectPokemon/SelectPokemonLike';
 import SelectPokemonRandom from '@/components/cardMaker/selectPokemon/SelectPokemonRandom';
-import styles from './cards.module.scss';
+import styles from './cardEditPage.module.scss';
 import PokemonCard from '../card/PokemonCard';
 import CardEditor from './CardEditor';
 import Inner from '../Inner';
