+++ src/components/cardMaker/cardEditPage.module.scss
@@ -1,0 +1,69 @@
+.card_page_wrapper {
+  width: 100%;
+  display: flex;
+  gap: 40px;
+  flex-direction: column;
+  align-items: center;
+}
+
+.product_card_wrapper {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+
+  > span {
+    margin-bottom: 20px;
+  }
+}
+
+.product_card_container_decoration {
+  &__top {
+    width: 100%;
+    height: 21px;
+    border-radius: 15px 15px 0px 0px;
+    border: 3px solid #000;
+    background: #ff5050;
+  }
+
+  &__bottom {
+    height: 21px;
+    border-radius: 0px 0px 15px 15px;
+    border: 3px solid #000;
+    background: #0f60ce;
+  }
+}
+
+.product_card_container {
+  align-items: center;
+  display: flex;
+  justify-content: center;
+  gap: 10%;
+  height: 588px;
+  border-left: 3px solid #000;
+  border-right: 3px solid #000;
+  background: #fff;
+}
+
+.select_wrapper {
+  width: 100%;
+  justify-content: space-between;
+  display: flex;
+}
+
+.title {
+  font-size: 28px;
+  font-weight: 700;
+}
+
+.save_button {
+  color: #fff;
+  text-align: center;
+  font-size: 21px;
+  font-weight: 500;
+  width: 280px;
+  height: 60px;
+  border-radius: 8px;
+  background-color: #0f60ce;
+  border: none;
+  cursor: pointer;
+}
